feat:general protocol support
