+++ .github/workflows/maven-publish.yml
@@ -28,10 +28,10 @@ jobs:
           gpg-passphrase: MAVEN_GPG_PASSPHRASE # env variable for GPG private key passphrase
  
       - name: Build with Maven
-        run: mvn -B -ntp clean package -DskipTests=true -Pjar
+        run: mvn -B -ntp clean package -DskipTests=true -pl arex-schedule-web-api
  
       - name: Publish to Apache Maven Central
-        run: mvn deploy -DskipTests=true -Pjar
+        run: mvn deploy -DskipTests=true -pl arex-schedule-web-api
         env:
           MAVEN_USERNAME: ${{ secrets.MAVEN_OSSRH_USERNAME }}
           MAVEN_CENTRAL_TOKEN: ${{ secrets.MAVEN_OSSRH_TOKEN }}
