+++ arex-schedule-extension/pom.xml
@@ -22,6 +22,16 @@
     </dependencies>
     <build>
         <finalName>arex-schedule-extension</finalName>
+        <plugins>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-compiler-plugin</artifactId>
+                <configuration>
+                    <source>8</source>
+                    <target>8</target>
+                </configuration>
+            </plugin>
+        </plugins>
     </build>
 
 </project>
