+++ arex-schedule-extension/src/main/java/com/arextest/schedule/extension/invoker/InvokerConstants.java
@@ -3,11 +3,11 @@ package com.arextest.schedule.extension.invoker;
 public class InvokerConstants {
     public static final String HEADERS = "headers";
 
-    //dubbo
-    public static final String INTERFACE_NAME = "interfaceName";
-    public static final String METHOD_NAME = "methodName";
-    public static final String PARAMETER_TYPES = "parameterTypes";
-    public static final String PARAMETERS = "parameters";
+    // dubbo
+    public static final String DUBBO_INTERFACE_NAME = "interfaceName";
+    public static final String DUBBO_METHOD_NAME = "methodName";
+    public static final String DUBBO_PARAMETER_TYPES = "parameterTypes";
+    public static final String DUBBO_PARAMETERS = "parameters";
 
     public static final String DUBBO_CASE_TYPE = "DubboProvider";
