+++ arex-schedule-extension/src/main/java/com/arextest/schedule/extension/invoker/ReplayExtensionInvoker.java
@@ -3,6 +3,9 @@ package com.arextest.schedule.extension.invoker;
 import com.arextest.schedule.extension.model.ReplayInvokeResult;
 
 public interface ReplayExtensionInvoker {
+    /**
+     * check caseType by InvokerConstants#XXXCaseType.
+     */
     boolean isSupported(String caseType);
 
     ReplayInvokeResult invoke(ReplayInvocation invocation);
