+++ arex-schedule-web-api/pom.xml
@@ -105,7 +105,7 @@
         <dependency>
             <groupId>com.arextest</groupId>
             <artifactId>arex-schedule-extension</artifactId>
-            <version>1.0.25.2</version>
+            <version>1.0.28</version>
             <scope>compile</scope>
         </dependency>
     </dependencies>
